Modify testsuite cases.
case broken becuase of PR tree-optimization/83648
Add option -flag_delete_null_pointer_checks
testsuite/gcc.dg/ipa/propmalloc-4.c

Add option -mno-riscv-iprintf:
testsuite/gcc.dg/tree-ssa/builtin-snprintf-2.c
testsuite/gcc.dg/tree-ssa/builtin-snprintf-4.c
testsuite/gcc.dg/tree-ssa/builtin-sprintf-6.c
testsuite/gcc.dg/tree-ssa/builtin-sprintf-8.c
testsuite/gcc.dg/tree-ssa/builtin-sprintf-warn-21.c
testsuite/gcc.dg/tree-ssa/pr79352.c
testsuite/gcc.dg/tree-ssa/pr79691.c
testsuite/gcc.dg/tree-ssa/builtin-snprintf-warn-4.c
testsuite/gcc.dg/pr47917.c
testsuite/gcc.dg/pr79538.c
testsuite/g++.dg/ext/attr-optimize.C

Fix dg-error format:
gcc/testsuite/gcc.target/riscv/nds-cpu-option-1.c
testsuite/gcc.target/riscv/nds-cpu-option-2.c

Delete builtin __riscv_zk.
testsuite/gcc.target/riscv/predef-18.c

Delete -misa-spec=2.2
gcc/testsuite/gcc.target/riscv/predef-10.c
gcc/testsuite/gcc.target/riscv/predef-11.c
gcc/testsuite/gcc.target/riscv/predef-13.c

Delete --help=target option because 'Undocumented' is wrong.
help.exp

Fix type casting to pointer type.
gcc/testsuite/gcc.target/riscv/cmov-3.c

Fix incompatible pointer type.
gcc/testsuite/gcc.target/riscv/fp16-compile-exprtype.c

Add -mabi.
gcc/testsuite/gcc.target/riscv/predef-14.c
gcc/testsuite/gcc.target/riscv/predef-15.c
gcc/testsuite/gcc.target/riscv/predef-16.c
gcc/testsuite/gcc.target/riscv/predef-17.c

+++ gcc/testsuite/gcc.target/riscv/cmov-3.c
@@ -2,24 +2,24 @@
 /* { dg-options "-O2 -mcmov" } */
 
 void *
-test1 (void *base, int sp, unsigned size)
+test1 (void *base, int sp, unsigned long size)
 {
   void *ret;
   for (int i = 0; i < 9; i++)
     {
-      ret = (void *) ((unsigned) base + size);
+      ret = (void *) ((unsigned long) base + size);
       size = 180 * (sp == 44100 ? 48000 : sp);
     }
   return ret;
 }
 
 void *
-test2 (void *base, int sp, unsigned size)
+test2 (void *base, int sp, unsigned long size)
 {
   void *ret;
   for (int i = 0; i < 9; i++)
     {
-      ret = (void *) ((unsigned) base + size);
+      ret = (void *) ((unsigned long) base + size);
       size = 180 * (sp != 44100 ? 48000 : sp);
     }
   return ret;
